fix: failure rate schema

+++ apps/api/src/app/metrics/resolvers/dora-metrics.schema.ts
@@ -112,7 +112,7 @@ export default /* GraphQL */ `
     columns: [DateTime!]!
 
     "The amounts over time for the chart"
-    data: [BigInt!]!
+    data: [Float!]!
   }
 
   type DeploymentFrequencyMetric {
