+++ packages/graphql-types/frontend/graphql.ts
@@ -193,7 +193,7 @@ export type ChangeFailureRateMetric = {
   /** The date range for the current period */
   currentPeriod: DateTimeRangeValue;
   /** The amounts over time for the chart */
-  data: Array<Scalars['BigInt']['output']>;
+  data: Array<Scalars['Float']['output']>;
   /** The change failure rate before the current period */
   previousAmount: Scalars['Float']['output'];
   /** The date range for the previous period */
